requirements.txt
streamlit 
langchain-google-genai  
pymupdf
python-dotenv
apify-client